backend/requirements.txt
# FastAPI and ASGI Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
geoalchemy2==0.14.2
psycopg2-binary==2.9.9
alembic==1.12.1
numpy<2  # Pin to 1.x for Shapely compatibility

# Redis
redis==5.0.1
# hiredis==2.2.3  # Commented out - requires Visual C++ Build Tools on Windows (optional optimization)

# Authentication
firebase-admin==6.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Image Processing
Pillow==10.1.0
python-magic==0.4.27

# WebSocket
python-socketio==5.10.0
websockets==12.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# Monitoring and Logging
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Task Queue (for async operations)
celery==5.3.4
flower==2.0.1

# Geospatial
shapely==2.0.2
geopy==2.4.1
polyline==2.0.0
pyproj==3.6.1

# Security
cryptography==41.0.7
bcrypt==4.1.1

# Rate Limiting
slowapi==0.1.9

# Environment
environs==10.0.0

# Geohashing
pygeohash==1.2.0

